Tweaks
Changed melee attack range logic, it should but much more accurate now,
this should also stop melee mobs from getting 'stuck' where they just
stand there and do nothing.

Geonach will now randomly charge towards their enemy breaking
surrounding blocks!

+++ java/lycanite/lycanitesmobs/api/spawning/SpawnTypeRock.java
@@ -59,7 +59,7 @@ public class SpawnTypeRock extends SpawnTypeBase {
     @Override
     public void spawnEntity(World world, EntityLiving entityLiving) {
         super.spawnEntity(world, entityLiving);
-        if(entityLiving instanceof EntityCreatureBase) {
+        if(entityLiving instanceof EntityCreatureBase && this.blockBreakRadius > -1) {
         	((EntityCreatureBase)entityLiving).destroyArea((int)entityLiving.posX, (int)entityLiving.posY, (int)entityLiving.posZ, 4, true, this.blockBreakRadius);
         }
     }
